// src/apiarys/dto/settings/update-settings.dto.ts
export class updateSettingsDto {

        apiaryId:number;
        apiaryUserId:number;
        honey?: boolean ;
        levudex?: boolean;
        sugar?: boolean;
        box?: boolean;
        boxMedium?: boolean;
        boxSmall?: boolean;
        tOxalic?: boolean;
        tAmitraz?: boolean;
        tFlumetrine?: boolean;
        tFence?: boolean;
        tComment?: boolean;
}